ISSUE #320 another attempt to remove ambiguity

+++ bouncer/src/repo/core/model/bson/repo_bson_factory.cpp
@@ -640,13 +640,13 @@ RepoRef RepoBSONFactory::makeRepoRef(
 	const std::string &fileName,
 	const RepoRef::RefType &type,
 	const std::string &link,
-	const uint64_t size) {
+	const uint32_t size) {
 
 	repo::core::model::RepoBSONBuilder builder;
 	builder.append(REPO_LABEL_ID, fileName);
 	builder.append(REPO_REF_LABEL_TYPE, RepoRef::convertTypeAsString(type));
 	builder.append(REPO_REF_LABEL_LINK, link);
-	builder.append(REPO_REF_LABEL_SIZE, size);
+	builder.append(REPO_REF_LABEL_SIZE, (unsigned int) size);
 	return RepoRef(builder.obj());
 }
